+++ cmd/influx/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -14,17 +15,15 @@ import (
 	"time"
 
 	"github.com/influxdata/influxdb/v2"
-	"github.com/influxdata/influxdb/v2/bolt"
 	"github.com/influxdata/influxdb/v2/cmd/influx/config"
 	"github.com/influxdata/influxdb/v2/cmd/influx/internal"
 	"github.com/influxdata/influxdb/v2/http"
 	"github.com/influxdata/influxdb/v2/internal/fs"
 	"github.com/influxdata/influxdb/v2/kit/cli"
-	"github.com/influxdata/influxdb/v2/kv"
 	"github.com/influxdata/influxdb/v2/pkg/httpc"
+	"github.com/influxdata/influxdb/v2/task/options"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"go.uber.org/zap"
 )
 
 const maxTCPConnections = 10
@@ -149,7 +148,6 @@ func runEMiddlware(mw cobraRunEMiddleware) genericCLIOptFn {
 
 type globalFlags struct {
 	config.Config
-	local        bool
 	skipVerify   bool
 	traceDebugID string
 }
@@ -236,7 +234,6 @@ func (b *cmdInfluxBuilder) cmd(childCmdFns ...func(f *globalFlags, opt genericCL
 	}
 	flags.Config = cfg
 
-	cmd.PersistentFlags().BoolVar(&flags.local, "local", false, "Run commands locally against the filesystem")
 	cmd.PersistentFlags().BoolVar(&flags.skipVerify, "skip-verify", false, "SkipVerify controls whether a client verifies the server's certificate chain and host name.")
 
 	// Update help description for all commands in command tree
@@ -413,20 +410,6 @@ func walk(c *cobra.Command, f func(*cobra.Command)) {
 	}
 }
 
-func newLocalKVService() (*kv.Service, error) {
-	boltFile, err := fs.BoltFile()
-	if err != nil {
-		return nil, err
-	}
-
-	store := bolt.NewKVStore(zap.NewNop(), boltFile)
-	if err := store.Open(context.Background()); err != nil {
-		return nil, err
-	}
-
-	return kv.NewService(zap.NewNop(), store), nil
-}
-
 type organization struct {
 	id, name string
 }
@@ -545,10 +528,6 @@ func writeJSON(w io.Writer, v interface{}) error {
 }
 
 func newBucketService() (influxdb.BucketService, error) {
-	if flags.local {
-		return newLocalKVService()
-	}
-
 	client, err := newHTTPClient()
 	if err != nil {
 		return nil, err
@@ -557,4 +536,51 @@ func newBucketService() (influxdb.BucketService, error) {
 	return &http.BucketService{
 		Client: client,
 	}, nil
+}
+
+func rawDurationToTimeDuration(raw string) (time.Duration, error) {
+	if raw == "" {
+		return 0, nil
+	}
+
+	if dur, err := time.ParseDuration(raw); err == nil {
+		return dur, nil
+	}
+
+	retention, err := options.ParseSignedDuration(raw)
+	if err != nil {
+		return 0, err
+	}
+
+	const (
+		day  = 24 * time.Hour
+		week = 7 * day
+	)
+
+	var dur time.Duration
+	for _, d := range retention.Values {
+		if d.Magnitude < 0 {
+			return 0, errors.New("must be greater than 0")
+		}
+		mag := time.Duration(d.Magnitude)
+		switch d.Unit {
+		case "w":
+			dur += mag * week
+		case "d":
+			dur += mag * day
+		case "m":
+			dur += mag * time.Minute
+		case "s":
+			dur += mag * time.Second
+		case "ms":
+			dur += mag * time.Minute
+		case "us":
+			dur += mag * time.Microsecond
+		case "ns":
+			dur += mag * time.Nanosecond
+		default:
+			return 0, errors.New("duration must be week(w), day(d), hour(h), min(m), sec(s), millisec(ms), microsec(us), or nanosec(ns)")
+		}
+	}
+	return dur, nil
 }