+++ cmd/influxd/launcher/pkger_test.go
@@ -183,10 +183,11 @@ func TestLauncher_Pkger(t *testing.T) {
 		return obj
 	}
 
-	newVariableObject := func(pkgName, name, description string) pkger.Object {
+	newVariableObject := func(pkgName, name, description string, selected ...string) pkger.Object {
 		obj := pkger.VariableToObject("", influxdb.Variable{
 			Name:        name,
 			Description: description,
+			Selected:    selected,
 			Arguments: &influxdb.VariableArguments{
 				Type:   "constant",
 				Values: influxdb.VariableConstantValues{"a", "b"},
@@ -1538,6 +1539,33 @@ func TestLauncher_Pkger(t *testing.T) {
 			})
 		})
 
+		t.Run("applying updates to existing variable should be successful", func(t *testing.T) {
+			stack, cleanup := newStackFn(t, pkger.Stack{})
+			defer cleanup()
+
+			impact, err := svc.Apply(ctx, l.Org.ID, l.User.ID,
+				pkger.ApplyWithStackID(stack.ID),
+				pkger.ApplyWithPkg(newPkg(newVariableObject("var", "", ""))),
+			)
+			require.NoError(t, err)
+
+			vars := impact.Summary.Variables
+			require.Len(t, vars, 1)
+			v := resourceCheck.mustGetVariable(t, byID(influxdb.ID(vars[0].ID)))
+			assert.Empty(t, v.Selected)
+
+			impact, err = svc.Apply(ctx, l.Org.ID, l.User.ID,
+				pkger.ApplyWithStackID(stack.ID),
+				pkger.ApplyWithPkg(newPkg(newVariableObject("var", "", "", "selected"))),
+			)
+			require.NoError(t, err)
+
+			vars = impact.Summary.Variables
+			require.Len(t, vars, 1)
+			v = resourceCheck.mustGetVariable(t, byID(influxdb.ID(vars[0].ID)))
+			assert.Equal(t, []string{"selected"}, v.Selected)
+		})
+
 		t.Run("apply with actions", func(t *testing.T) {
 			var (
 				bucketPkgName   = "rucketeer-1"
@@ -1551,12 +1579,29 @@ func TestLauncher_Pkger(t *testing.T) {
 				variablePkgName = "laces-out-dan"
 			)
 
+			defaultPkgFn := func(*testing.T) *pkger.Pkg {
+				return newPkg(
+					newBucketObject(bucketPkgName, "", ""),
+					newCheckDeadmanObject(t, checkPkgName, "", time.Hour),
+					newDashObject(dashPkgName, "", ""),
+					newEndpointHTTP(endpointPkgName, "", ""),
+					newLabelObject(labelPkgName, "", "", ""),
+					newRuleObject(t, rulePkgName, "", endpointPkgName, ""),
+					newTaskObject(taskPkgName, "", ""),
+					newTelegrafObject(telegrafPkgName, "", ""),
+					newVariableObject(variablePkgName, "", ""),
+				)
+			}
+
 			tests := []struct {
 				name      string
+				pkgFn     func(t *testing.T) *pkger.Pkg
 				applyOpts []pkger.ApplyOptFn
+				assertFn  func(t *testing.T, impact pkger.PkgImpactSummary)
 			}{
 				{
-					name: "skip resource",
+					name:  "skip resource",
+					pkgFn: defaultPkgFn,
 					applyOpts: []pkger.ApplyOptFn{
 						pkger.ApplyWithResourceSkip(pkger.ActionSkipResource{
 							Kind:     pkger.KindBucket,
@@ -1595,9 +1640,22 @@ func TestLauncher_Pkger(t *testing.T) {
 							MetaName: variablePkgName,
 						}),
 					},
+					assertFn: func(t *testing.T, impact pkger.PkgImpactSummary) {
+						summary := impact.Summary
+						assert.Empty(t, summary.Buckets)
+						assert.Empty(t, summary.Checks)
+						assert.Empty(t, summary.Dashboards)
+						assert.Empty(t, summary.NotificationEndpoints)
+						assert.Empty(t, summary.Labels)
+						assert.Empty(t, summary.NotificationRules, 0)
+						assert.Empty(t, summary.Tasks)
+						assert.Empty(t, summary.TelegrafConfigs)
+						assert.Empty(t, summary.Variables)
+					},
 				},
 				{
-					name: "skip kind",
+					name:  "skip kind",
+					pkgFn: defaultPkgFn,
 					applyOpts: []pkger.ApplyOptFn{
 						pkger.ApplyWithKindSkip(pkger.ActionSkipKind{
 							Kind: pkger.KindBucket,
@@ -1627,6 +1685,70 @@ func TestLauncher_Pkger(t *testing.T) {
 							Kind: pkger.KindVariable,
 						}),
 					},
+					assertFn: func(t *testing.T, impact pkger.PkgImpactSummary) {
+						summary := impact.Summary
+						assert.Empty(t, summary.Buckets)
+						assert.Empty(t, summary.Checks)
+						assert.Empty(t, summary.Dashboards)
+						assert.Empty(t, summary.NotificationEndpoints)
+						assert.Empty(t, summary.Labels)
+						assert.Empty(t, summary.NotificationRules, 0)
+						assert.Empty(t, summary.Tasks)
+						assert.Empty(t, summary.TelegrafConfigs)
+						assert.Empty(t, summary.Variables)
+					},
+				},
+				{
+					name: "skip label and assoications should be dropped",
+					pkgFn: func(t *testing.T) *pkger.Pkg {
+						objs := []pkger.Object{
+							newBucketObject(bucketPkgName, "", ""),
+							newCheckDeadmanObject(t, checkPkgName, "", time.Hour),
+							newDashObject(dashPkgName, "", ""),
+							newEndpointHTTP(endpointPkgName, "", ""),
+							newRuleObject(t, rulePkgName, "", endpointPkgName, ""),
+							newTaskObject(taskPkgName, "", ""),
+							newTelegrafObject(telegrafPkgName, "", ""),
+							newVariableObject(variablePkgName, "", ""),
+						}
+						for _, obj := range objs {
+							obj.AddAssociations(pkger.ObjectAssociation{
+								Kind:    pkger.KindLabel,
+								PkgName: labelPkgName,
+							})
+						}
+
+						objs = append(objs, newLabelObject(labelPkgName, "", "", ""))
+
+						return newPkg(objs...)
+					},
+					applyOpts: []pkger.ApplyOptFn{
+						pkger.ApplyWithKindSkip(pkger.ActionSkipKind{
+							Kind: pkger.KindLabel,
+						}),
+					},
+					assertFn: func(t *testing.T, impact pkger.PkgImpactSummary) {
+						summary := impact.Summary
+						assert.Empty(t, summary.Labels, 0)
+						assert.Empty(t, summary.LabelMappings)
+
+						assert.Len(t, summary.Buckets, 1)
+						assert.Empty(t, summary.Buckets[0].LabelAssociations)
+						assert.Len(t, summary.Checks, 1)
+						assert.Empty(t, summary.Checks[0].LabelAssociations)
+						assert.Len(t, summary.Dashboards, 1)
+						assert.Empty(t, summary.Dashboards[0].LabelAssociations)
+						assert.Len(t, summary.NotificationEndpoints, 1)
+						assert.Empty(t, summary.NotificationEndpoints[0].LabelAssociations)
+						assert.Len(t, summary.NotificationRules, 1)
+						assert.Empty(t, summary.NotificationRules[0].LabelAssociations)
+						assert.Len(t, summary.Tasks, 1)
+						assert.Empty(t, summary.Tasks[0].LabelAssociations)
+						assert.Len(t, summary.TelegrafConfigs, 1)
+						assert.Empty(t, summary.TelegrafConfigs[0].LabelAssociations)
+						assert.Len(t, summary.Variables, 1)
+						assert.Empty(t, summary.Variables[0].LabelAssociations)
+					},
 				},
 			}
 
@@ -1635,37 +1757,16 @@ func TestLauncher_Pkger(t *testing.T) {
 					stack, cleanup := newStackFn(t, pkger.Stack{})
 					defer cleanup()
 
-					pkg := newPkg(
-						newBucketObject(bucketPkgName, "", ""),
-						newCheckDeadmanObject(t, checkPkgName, "", time.Hour),
-						newDashObject(dashPkgName, "", ""),
-						newEndpointHTTP(endpointPkgName, "", ""),
-						newLabelObject(labelPkgName, "", "", ""),
-						newRuleObject(t, rulePkgName, "", endpointPkgName, ""),
-						newTaskObject(taskPkgName, "", ""),
-						newTelegrafObject(telegrafPkgName, "", ""),
-						newVariableObject(variablePkgName, "", ""),
-					)
-
 					impact, err := svc.Apply(ctx, l.Org.ID, l.User.ID,
 						append(
 							tt.applyOpts,
-							pkger.ApplyWithPkg(pkg),
+							pkger.ApplyWithPkg(tt.pkgFn(t)),
 							pkger.ApplyWithStackID(stack.ID),
 						)...,
 					)
 					require.NoError(t, err)
 
-					summary := impact.Summary
-					assert.Empty(t, summary.Buckets)
-					assert.Empty(t, summary.Checks)
-					assert.Empty(t, summary.Dashboards)
-					assert.Empty(t, summary.NotificationEndpoints)
-					assert.Empty(t, summary.Labels)
-					assert.Empty(t, summary.NotificationRules, 0)
-					assert.Empty(t, summary.Tasks)
-					assert.Empty(t, summary.TelegrafConfigs)
-					assert.Empty(t, summary.Variables)
+					tt.assertFn(t, impact)
 				}
 
 				t.Run(tt.name, fn)
@@ -2310,6 +2411,7 @@ spec:
 		require.Len(t, vars, 1)
 		assert.NotZero(t, vars[0].ID)
 		assert.Equal(t, "query var", vars[0].Name)
+		assert.Equal(t, []string{"rucketeer"}, vars[0].Selected)
 		hasLabelAssociations(t, vars[0].LabelAssociations, 1, "label-1")
 		varArgs := vars[0].Arguments
 		require.NotNil(t, varArgs)
@@ -2318,6 +2420,8 @@ spec:
 			Query:    "buckets()  |> filter(fn: (r) => r.name !~ /^_/)  |> rename(columns: {name: \"_value\"})  |> keep(columns: [\"_value\"])",
 			Language: "flux",
 		}, varArgs.Values)
+		platformVar := resourceCheck.mustGetVariable(t, byID(influxdb.ID(vars[0].ID)))
+		assert.Equal(t, []string{"rucketeer"}, platformVar.Selected)
 
 		newSumMapping := func(id pkger.SafeID, pkgName, name string, rt influxdb.ResourceType) pkger.SummaryLabelMapping {
 			return pkger.SummaryLabelMapping{
@@ -3149,6 +3253,8 @@ spec:
   language: flux
   query: |
     buckets()  |> filter(fn: (r) => r.name !~ /^_/)  |> rename(columns: {name: "_value"})  |> keep(columns: ["_value"])
+  selected:
+    - rucketeer
   associations:
     - kind: Label
       name: label-1