+++ pkger/parser_test.go
@@ -60,12 +60,12 @@ func TestParse(t *testing.T) {
 					{
 						Field:        "metadata.name",
 						EnvRefKey:    "meta-name",
-						DefaultValue: "env-meta-name",
+						DefaultValue: "meta",
 					},
 					{
 						Field:        "spec.name",
 						EnvRefKey:    "spec-name",
-						DefaultValue: "env-spec-name",
+						DefaultValue: "spectacles",
 					},
 					{
 						Field:        "spec.associations[0].name",
@@ -582,12 +582,12 @@ spec:
 					{
 						Field:        "metadata.name",
 						EnvRefKey:    "meta-name",
-						DefaultValue: "env-meta-name",
+						DefaultValue: "meta",
 					},
 					{
 						Field:        "spec.name",
 						EnvRefKey:    "spec-name",
-						DefaultValue: "env-spec-name",
+						DefaultValue: "spectacles",
 					},
 					{
 						Field:        "spec.associations[0].name",
@@ -2281,12 +2281,12 @@ spec:
 					{
 						Field:        "metadata.name",
 						EnvRefKey:    "meta-name",
-						DefaultValue: "env-meta-name",
+						DefaultValue: "meta",
 					},
 					{
 						Field:        "spec.name",
 						EnvRefKey:    "spec-name",
-						DefaultValue: "env-spec-name",
+						DefaultValue: "spectacles",
 					},
 					{
 						Field:        "spec.associations[0].name",
@@ -2528,12 +2528,12 @@ spec:
 					{
 						Field:        "metadata.name",
 						EnvRefKey:    "meta-name",
-						DefaultValue: "env-meta-name",
+						DefaultValue: "meta",
 					},
 					{
 						Field:        "spec.name",
 						EnvRefKey:    "spec-name",
-						DefaultValue: "env-spec-name",
+						DefaultValue: "spectacles",
 					},
 					{
 						Field:        "spec.associations[0].name",
@@ -2852,12 +2852,12 @@ spec:
 					{
 						Field:        "metadata.name",
 						EnvRefKey:    "meta-name",
-						DefaultValue: "env-meta-name",
+						DefaultValue: "meta",
 					},
 					{
 						Field:        "spec.name",
 						EnvRefKey:    "spec-name",
-						DefaultValue: "env-spec-name",
+						DefaultValue: "spectacles",
 					},
 					{
 						Field:        "spec.associations[0].name",
@@ -3189,12 +3189,12 @@ spec:
 					{
 						Field:        "metadata.name",
 						EnvRefKey:    "meta-name",
-						DefaultValue: "env-meta-name",
+						DefaultValue: "meta",
 					},
 					{
 						Field:        "spec.name",
 						EnvRefKey:    "spec-name",
-						DefaultValue: "env-spec-name",
+						DefaultValue: "spectacles",
 					},
 					{
 						Field:        "spec.associations[0].name",
@@ -3407,12 +3407,12 @@ spec:
 					{
 						Field:        "metadata.name",
 						EnvRefKey:    "meta-name",
-						DefaultValue: "env-meta-name",
+						DefaultValue: "meta",
 					},
 					{
 						Field:        "spec.name",
 						EnvRefKey:    "spec-name",
-						DefaultValue: "env-spec-name",
+						DefaultValue: "spectacles",
 					},
 					{
 						Field:        "spec.associations[0].name",
@@ -3471,11 +3471,15 @@ spec:
 
 				require.Len(t, sum.Variables, 4)
 
-				varEquals := func(t *testing.T, name, vType string, vals interface{}, v SummaryVariable) {
+				varEquals := func(t *testing.T, name, vType string, vals interface{}, selected []string, v SummaryVariable) {
 					t.Helper()
 
 					assert.Equal(t, name, v.Name)
 					assert.Equal(t, name+" desc", v.Description)
+					if selected == nil {
+						selected = []string{}
+					}
+					assert.Equal(t, selected, v.Selected)
 					require.NotNil(t, v.Arguments)
 					assert.Equal(t, vType, v.Arguments.Type)
 					assert.Equal(t, vals, v.Arguments.Values)
@@ -3486,6 +3490,7 @@ spec:
 					"var-const-3",
 					"constant",
 					influxdb.VariableConstantValues([]string{"first val"}),
+					nil,
 					sum.Variables[0],
 				)
 
@@ -3493,6 +3498,7 @@ spec:
 					"var-map-4",
 					"map",
 					influxdb.VariableMapValues{"k1": "v1"},
+					nil,
 					sum.Variables[1],
 				)
 
@@ -3503,6 +3509,7 @@ spec:
 						Query:    `buckets()  |> filter(fn: (r) => r.name !~ /^_/)  |> rename(columns: {name: "_value"})  |> keep(columns: ["_value"])`,
 						Language: "flux",
 					},
+					[]string{"rucket"},
 					sum.Variables[2],
 				)
 
@@ -3513,6 +3520,7 @@ spec:
 						Query:    "an influxql query of sorts",
 						Language: "influxql",
 					},
+					nil,
 					sum.Variables[3],
 				)
 			})
@@ -3527,17 +3535,27 @@ spec:
 					{
 						Field:        "metadata.name",
 						EnvRefKey:    "meta-name",
-						DefaultValue: "env-meta-name",
+						DefaultValue: "meta",
 					},
 					{
 						Field:        "spec.name",
 						EnvRefKey:    "spec-name",
-						DefaultValue: "env-spec-name",
+						DefaultValue: "spectacles",
 					},
 					{
 						Field:        "spec.associations[0].name",
 						EnvRefKey:    "label-meta-name",
 						DefaultValue: "env-label-meta-name",
+					},
+					{
+						Field:        "spec.selected[0]",
+						EnvRefKey:    "the-selected",
+						DefaultValue: "second val",
+					},
+					{
+						Field:        "spec.selected[1]",
+						EnvRefKey:    "the-2nd",
+						DefaultValue: "env-the-2nd",
 					},
 				}
 				assert.Equal(t, expectedEnvRefs, actual[0].EnvReferences)